Refactor - moved all service routines from ftClient.cpp to their own modules
